Refactor auth context, update imports, and add Supabase env example
