+++ .env.example
@@ -1,0 +1,10 @@
+# Supabase Configuration
+# Copy this file to .env and fill in your actual Supabase project details
+# You can find these values in your Supabase project settings
+
+EXPO_PUBLIC_SUPABASE_URL=your_supabase_project_url_here
+EXPO_PUBLIC_SUPABASE_ANON_KEY=your_supabase_anon_key_here
+
+# Example:
+# EXPO_PUBLIC_SUPABASE_URL=https://your-project.supabase.co
+# EXPO_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
