+++ hooks/useFrameworkReady.ts
@@ -1,4 +1,5 @@
 import { useEffect } from 'react';
+import { Platform } from 'react-native';
 
 declare global {
   interface Window {
@@ -8,6 +9,9 @@ declare global {
 
 export function useFrameworkReady() {
   useEffect(() => {
-    window.frameworkReady?.()
-  }, [])
+    // Only call frameworkReady on web platform
+    if (Platform.OS === 'web') {
+      (window as any)?.frameworkReady?.();
+    }
+  }, []);
 }
